requirements.txt
fastapi
uvicorn
twilio
llama-index
llama-index-llms-gemini
python-dotenv
chromadb
sentence-transformers
pandas
